Refactor API calls into a separate module
In advance of introducing calls to the unofficial API in a separate
module.  Separate modules makes it easier to keep things straight once
there are two kinds of GET, POST, etc requests.
