+++ src/readthedocs_cli/api/__init__.py
@@ -1,0 +1,1 @@
+from . import v3
